coreinit: Fix warning about implicit size_t conversion to int.

+++ src/libdecaf/src/modules/coreinit/coreinit_sprintf.cpp
@@ -184,7 +184,7 @@ formatStringV(char *buffer,
       buffer[str.length()] = 0;
    }
 
-   return str.length();
+   return static_cast<int>(str.length());
 }
 
 } // namespace internal
